Split models into multiple files and update structure of results
This is a fairly major update to the creel portal models, and as a
result most of the views do not currently work. Updates include
splitting the models file into three separate files for design tables,
the fishnet tables and the results tables. Additionally a
slug field has been added to most tables, and the relationship between
results tables now more explicit and follow the original structure of
creesys much more closely.  Next step - set up testing to verify
everything works as it should.