Updated tests and factories to reflect model structure
This commit updates all of the existing tests and factories to mirror
the file structure inside of ./models.  All of the existing tests pass,
except for the first initial test of the api (which has not been built
yet). The old files that housed all of the factories and tests have
been removed.