Creel detail view is working again
After updating the models, the creel detail view stopped working. This
commit includes changes to get it working again, including moving some
model methods to their own utility functions to avoid circular
imports.  The template was also updated to handle the situation where
creel results are not currently available and provide a meaningful
message when that is the case.  Still need to write unit tests for the
creel portal.