Fixed the catch and effort estimates views
The details that display the catch and effort estimates with a map and
interactive graphics broke when we updated the models. This commit
includes changes to those templates and creates two new responses that
return data efficiently as json response.  Still more to do, but at
least things seem to work as they use to.