Added detail api endpoints and tests for creel tables FN011-FN028
This commit includes the addition of detail api views for each of the
creel design tables and associated test to verify that they return the
expected elements.  This commit also includes a permission class for
the api endpoints that allow administrators and project leads to
edit the creel design endpoints.