Added clean method and tests to FN025 model to verify date
Added a clean method to the FN025 model to ensure that it always falls
within the start and end date of its associated season. An error is
raised if we try to save an exception date outside of those bounds. An
associated unit is included to ensure that the functionality continues
to work as expected.