Added setup.py files, migrations, and updates to templates
This commit is a big one and includes a number of changes or additions
that will make creel_portal an installable application.  It includes
all of the required files (setup, manifest, readme...) as well as a
base template that is more explicitly associated with this project. It
also removed related_names from foreign key relationships back to
entities in the common data application to avoid clashes with other
applications.  Finally, it includes all migrations required up to this
point.  All of the existing tests pass.