Read-only list-view, filters, and serializer for FN111 objects.
This commit adds a read-only listview, filters, and serializer for
FN111 objects as well as refinements to the filters for the other
endpoints - links to related fields in the field argument cause
additional, unnecessary queries.  There are currently no tests
assocaited with the new code added in this commit, but there will be
shortly. All of the previously existing tests continue to pass.