Read-only api endpoint for FN123 - catch counts
This commit adds a fast and flat api endpoint that closely matches
the FN-2 data format. The commit includes a view, a filter, a
serializer and a url route.  All of the 'fast and flat' urls have been
changed from /fn* to /sc* to differentiate them from exiting /fn*
routes for fn_portal.  No tests are implemented for this code, but all
of the existing and previously passing tests continue to pass.