Read-only api endpoints for FN125 and FN125_Lamprey
This commit includes read-only api listviews for FN125 and
fn125Lamprey objects, as well as filters and serializers. A generic
FishAttr_filter has also been created that will be used for all
endpoints that are children of FN125.  No tests were included in this
commit, but all of the previously existing tests continue to pass.