Updated creel detail page to include counts and zoom-to-project
This commit updates the creel detail page to include additional
summary statistics - interview count, and biological samples. It also
add a couple of buttons and associated javascript to toggle the extent
of the map zoom between the extent of the lake, and the extent of the
creel. All previously passing tests continue to pass.